server/requirements.txt
aiohttp==3.8.1
pymupdf==1.19.6
aiosignal==1.2.0
anyascii==0.3.0
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
astor==0.8.1
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
Babel==2.9.1
bce-python-sdk==0.8.64
beautifulsoup4==4.10.0
Bottleneck==1.3.4
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
ci-info==0.2.0
click==8.0.4
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.6.0
configobj==5.0.6
configparser==5.2.0
coqpit==0.0.15
coqui-trainer==0.0.5
croniter==1.3.4
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.28
datasets==2.1.0
dateparser==1.1.1
decorator==4.4.2
Deprecated==1.2.13
dill==0.3.4
Distance==0.1.3
distlib==0.3.4
docopt==0.6.2
dtaidistance==2.3.1
editdistance==0.6.0
etelemetry==0.3.0
fastapi==0.75.0
feedparser==6.0.8
filelock==3.6.0
flake8==4.0.1
Flask==2.0.3
Flask-Babel==2.0.0
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
g2p-en==2.1.0
g2pM==0.1.2.5
gdown==4.4.0
greenlet==1.1.2
gruut==2.3.0
gruut-ipa==0.13.0
h11==0.13.0
h5py==3.6.0
httplib2==0.20.4
huggingface-hub==0.5.1
humanfriendly==10.0
identify==2.5.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
img2pdf==0.4.3
inflect==5.4.0
isodate==0.6.1
itsdangerous==2.1.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
jsonlines==1.2.0
kaldiio==2.17.2
kiwisolver==1.4.0
langdetect==1.0.9
librosa==0.8.1
llvmlite==0.36.0
loguru==0.6.0
looseversion==1.0.1
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib==3.5.1
mccabe==0.6.1
mecab-python3==1.0.3
minio==7.1.5
mock==4.0.3
moviepy==1.0.3
msgpack-python==0.5.6
multidict==6.0.2
multiprocess==0.70.12.2
nara-wpe==0.0.8
networkx==2.7.1
nibabel==3.2.2
nipype==1.8.1
nltk==3.7
nodeenv==1.6.0
num2words==0.5.10
numba==0.53.0
numpy==1.22.3
ocrmypdf==13.2.0
packaging==21.3
paddleaudio==0.2.1
paddlefsl==1.1.0
paddlenlp==2.2.6
paddlepaddle==2.2.2
paddlespeech==0.2.0
paddlespeech-feat==0.1.0
pandas==1.4.1
pathlib==1.0.1
pathos==0.2.8
pattern-singleton==1.2.0
pdf2image==1.16.0
pdfminer.six==20211012
pikepdf==5.0.1
Pillow==8.4.0
platformdirs==2.5.2
playwright==1.17.2
pluggy==1.0.0
pooch==1.6.0
portalocker==2.4.0
pox==0.3.0
ppft==1.6.6.4
praatio==5.0.0
pre-commit==2.19.0
prettytable==3.3.0
proglog==0.1.9
protobuf==3.19.4
prov==2.0.0
pyarrow==8.0.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pydot==1.4.2
pydotenv==0.0.7
pydub==0.25.1
pyee==9.0.4
pyflakes==2.4.0
PyJWT==2.3.0
PyMuPDF==1.19.6
PyMySQL==1.0.2
pynndescent==0.5.6
pyparsing==3.0.7
pypinyin==0.46.0
pypinyin-dict==0.2.0
PyRSS2Gen==1.1
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyworld==0.3.0
pyxnat==1.4
PyYAML==6.0
rdflib==6.1.1
redis==4.1.4
regex==2022.3.15
reportlab==3.6.8
requests==2.26.0
resampy==0.2.2
responses==0.18.0
rq==1.10.1
rq-scheduler==0.11.0
sacrebleu==2.0.0
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
sgmllib3k==1.0.0
shellcheck-py==0.8.0.4
simplejson==3.17.6
singleton-decorator==1.0.0
six==1.16.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
starlette-context==0.3.3
tabulate==0.8.9
tensorboardX==2.5
TextGrid==1.5
thefuzz==0.19.0
threadpoolctl==3.1.0
timer==0.2.2
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
tqdm==4.63.0
traits==6.3.2
TTS==0.5.0
typeguard==2.13.3
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.8
uvicorn==0.17.6
validators==0.18.2
virtualenv==20.14.1
visualdl==2.2.3
wcwidth==0.2.5
webrtcvad==2.0.10
websockets==10.2
Werkzeug==2.0.3
wrapt==1.14.0
xxhash==3.0.0
yacs==0.1.8
yarl==1.7.2
you-get==0.4.1602
fire==0.4.0
termcolor==2.0.1
fuzzysearch==0.7.3
